Make feature more general
The solution was overly specific, so we changed it to work for all 404s.
We ignore some paths that are never reachable (right now `tree` and `blob`).
For now it also doesn't change `blob` to `commits` anymore.
